Factor out derived entities tabs
