+++ context/app/static/js/components/Detail/DerivedEntitiesSection/DerivedEntitiesSection.jsx
@@ -1,21 +1,14 @@
-import React, { useState } from 'react';
+import React from 'react';
 import Button from '@material-ui/core/Button';
 import CircularProgress from '@material-ui/core/CircularProgress';
 
-import { Tabs, Tab } from 'js/shared-styles/tabs';
 import SectionHeader from 'js/shared-styles/sections/SectionHeader';
 import SectionContainer from 'js/shared-styles/sections/SectionContainer';
 import { StyledButtonRow, BottomAlignedTypography } from 'js/shared-styles/sections/RightAlignedButtonRow';
-import DerivedEntitiesTable from 'js/components/Detail/DerivedEntitiesTable';
-import { StyledDiv, StyledTabPanel, StyledCenteredLoaderWrapper } from './style';
+import DerivedEntitiesTabs from 'js/components/Detail/DerivedEntitiesTabs';
+import { StyledDiv, StyledCenteredLoaderWrapper } from './style';
 
 function DerivedEntitiesSection({ samples, datasets, uuid, isLoading, entityType }) {
-  const [openIndex, setOpenIndex] = useState(0);
-
-  const handleChange = (event, newIndex) => {
-    setOpenIndex(newIndex);
-  };
-
   return isLoading ? (
     <StyledCenteredLoaderWrapper>
       <CircularProgress />
@@ -43,20 +36,7 @@ function DerivedEntitiesSection({ samples, datasets, uuid, isLoading, entityType
             </Button>
           }
         />
-        <Tabs value={openIndex} onChange={handleChange} aria-label="Provenance Tabs">
-          {samples && <Tab label="Samples" index={0} />}
-          {datasets && <Tab label="Datasets" index={1} />}
-        </Tabs>
-        {samples && (
-          <StyledTabPanel value={openIndex} index={0}>
-            <DerivedEntitiesTable entities={samples} entityType="Sample" />
-          </StyledTabPanel>
-        )}
-        {datasets && (
-          <StyledTabPanel value={openIndex} index={1}>
-            <DerivedEntitiesTable entities={datasets} entityType="Dataset" />
-          </StyledTabPanel>
-        )}
+        <DerivedEntitiesTabs samples={samples} datasets={datasets} />
       </StyledDiv>
     </SectionContainer>
   );
