+++ context/app/static/js/components/Detail/DerivedEntitiesSection/style.js
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-import TabPanel from 'js/shared-styles/tabs/TabPanel';
 import CenteredLoaderWrapper from 'js/shared-styles/loaders/CenteredLoaderWrapper';
 
 const sectionHeight = 400; // px
@@ -10,12 +9,8 @@ const StyledDiv = styled.div`
   flex-direction: column;
 `;
 
-const StyledTabPanel = styled(TabPanel)`
-  flex-grow: 1;
-`;
-
 const StyledCenteredLoaderWrapper = styled(CenteredLoaderWrapper)`
   height: ${sectionHeight}px;
 `;
 
-export { StyledDiv, StyledTabPanel, StyledCenteredLoaderWrapper };
+export { StyledDiv, StyledCenteredLoaderWrapper };
