+++ context/app/static/js/components/Detail/DerivedEntitiesTabs/index.js
@@ -1,0 +1,3 @@
+import DerivedEntitiesTabs from './DerivedEntitiesTabs';
+
+export default DerivedEntitiesTabs;
