+++ context/app/static/js/components/Detail/DerivedEntitiesTabs/style.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import TabPanel from 'js/shared-styles/tabs/TabPanel';
+
+const StyledTabPanel = styled(TabPanel)`
+  flex-grow: 1;
+`;
+
+export { StyledTabPanel };
